+++ changelog.md
@@ -3,6 +3,12 @@ This file contains all the notable changes done to the Ballerina Email package t
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- [Changed disallowing service level annotations in the compiler plugin](https://github.com/ballerina-platform/ballerina-standard-library/issues/4741)
+
 ## [2.7.1] - 2023-06-01
 
 - [Fixed log manager getting reset by `ftp` module](https://github.com/ballerina-platform/ballerina-standard-library/issues/4478)